+++ src/commands.ts
@@ -80,6 +80,7 @@ import './commands/showQuickRepoStatus.js';
 import './commands/showQuickStashList.js';
 import './commands/showView.js';
 import './commands/signing/setup.js';
+import './commands/startAgentSession.js';
 import './commands/stashApply.js';
 import './commands/stashSave.js';
 import './commands/switchMode.js';